refactor(phase-3): split design_system & animation_helper into packages
- design_system.py (579 lines) → design_system/ package with 4 modules:
  - _tokens.py: Colors/Radius/Fonts/Motion/Shadows design tokens
  - _style_generator.py: StyleSheet class (QSS generator)
  - _components.py: CFButton/CFLabel/CFCard/CFPanel/CFInput/CFProgressBar/CFNavButton/CFToastNotification
  - __init__.py: re-exports for backward compat

- animation_helper.py (531 lines) → animation/ package with 4 modules:
  - _animation_helper.py: AnimationHelper class with platform-aware
    prefers-reduced-motion detection (macOS/Windows/Linux)
  - _page_transition.py: PageTransition class
  - _widgets.py: AnimatedButton & LoadingAnimation
  - __init__.py: re-exports for backward compat

- Update design_system/_components.py to use new animation import path

All modules now ≤303 lines (target was 300, _animation_helper is just
over due to platform detection logic)
Lint: ruff check all green
Tests: 46 passed, 1 skipped
