+++ src/scenefab/ui/theme/animation/_animation_helper.py
@@ -1,13 +1,19 @@
-"""
-动效工具类 - 提供各种动画效果
-支持 prefers-reduced-motion 系统偏好检测
-frontend-design-pro 规范 · 2026-04-13
-"""
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
 
+"""
+AnimationHelper — 通用动画辅助类
+
+提供：
+- prefers-reduced-motion 系统偏好检测（跨平台）
+- 淡入 / 淡出 / 滑入 / 缩放 / 脉冲 动画
+- 动画时长自动适配（减少动画时缩短至 10ms 内）
+"""
 import platform
 import logging
 from PySide6.QtWidgets import QWidget, QGraphicsOpacityEffect
-from PySide6.QtCore import QPropertyAnimation, QEasingCurve, QTimer, Qt, QSettings
+from PySide6.QtCore import QPropertyAnimation, QEasingCurve, QSettings
+
 logger = logging.getLogger(__name__)
 
 
@@ -294,238 +300,4 @@ class AnimationHelper:
         cls._reduced_motion_cache = None
 
 
-class PageTransition:
-    """页面切换动画管理器"""
-
-    def __init__(self, stacked_widget):
-        self.stacked_widget = stacked_widget
-
-    def switch_page(self, index: int, animation_type: str = "fade"):
-        """切换页面并播放动画
-
-        Args:
-            index: 目标页面索引
-            animation_type: 动画类型 ("fade", "slide", "scale")
-
-        当系统启用了 prefers-reduced-motion 时：
-        - 直接切换，无动画效果
-        """
-        if index < 0 or index >= self.stacked_widget.count():
-            return
-
-        # 如果启用 reduce motion，直接切换
-        if AnimationHelper._prefers_reduced_motion():
-            self.stacked_widget.setCurrentIndex(index)
-            return
-
-        current_widget = self.stacked_widget.currentWidget()
-        target_widget = self.stacked_widget.widget(index)
-
-        _TRANSITION_MAP = {
-            "fade": self._fade_transition,
-            "slide": self._slide_transition,
-            "scale": self._scale_transition,
-        }
-        transition = _TRANSITION_MAP.get(animation_type)
-        if transition:
-            transition(current_widget, target_widget)
-        else:
-            self.stacked_widget.setCurrentIndex(index)
-
-    def _fade_transition(self, from_widget: QWidget, to_widget: QWidget):
-        """淡入淡出切换"""
-        # 淡出当前页面
-        self.fade_out(from_widget, duration=200)
-
-        # 切换到目标页面并淡入
-        QTimer.singleShot(200, lambda: [
-            self.stacked_widget.setCurrentWidget(to_widget),
-            self.fade_in(to_widget, duration=200)
-        ])
-
-    def _slide_transition(self, from_widget: QWidget, to_widget: QWidget):
-        """滑动切换"""
-        # 获取宽度
-        _ = self.stacked_widget.width()  # needed to compute transition
-
-        # 淡出当前页面
-        self.fade_out(from_widget, duration=150)
-
-        # 切换并淡入目标页面
-        QTimer.singleShot(150, lambda: [
-            self.stacked_widget.setCurrentWidget(to_widget),
-            self.fade_in(to_widget, duration=150)
-        ])
-
-    def _scale_transition(self, from_widget: QWidget, to_widget: QWidget):
-        """缩放切换"""
-        self.scale_out(from_widget, duration=150)
-
-        QTimer.singleShot(150, lambda: [
-            self.stacked_widget.setCurrentWidget(to_widget),
-            self.scale_in(to_widget, duration=150)
-        ])
-
-    def fade_in(self, widget: QWidget, duration: int = 200):
-        """淡入"""
-        actual_duration = AnimationHelper._get_reduced_duration(duration)
-        if actual_duration < 10:
-            widget.show()
-            return
-
-        opacity = QGraphicsOpacityEffect(widget)
-        widget.setGraphicsEffect(opacity)
-
-        animation = QPropertyAnimation(opacity, b"opacity")
-        animation.setDuration(actual_duration)
-        animation.setStartValue(0)
-        animation.setEndValue(1)
-        animation.setEasingCurve(QEasingCurve.Type.InOutQuad)
-        animation.start()
-
-    def fade_out(self, widget: QWidget, duration: int = 200):
-        """淡出"""
-        actual_duration = AnimationHelper._get_reduced_duration(duration)
-        if actual_duration < 10:
-            widget.hide()
-            return
-
-        opacity = QGraphicsOpacityEffect(widget)
-        widget.setGraphicsEffect(opacity)
-
-        animation = QPropertyAnimation(opacity, b"opacity")
-        animation.setDuration(actual_duration)
-        animation.setStartValue(1)
-        animation.setEndValue(0)
-        animation.setEasingCurve(QEasingCurve.Type.InOutQuad)
-        animation.start()
-
-    def scale_in(self, widget: QWidget, duration: int = 150):
-        """缩放入场"""
-        actual_duration = AnimationHelper._get_reduced_duration(duration)
-        if actual_duration < 10:
-            widget.show()
-            return
-
-        opacity = QGraphicsOpacityEffect(widget)
-        widget.setGraphicsEffect(opacity)
-
-        # 透明度动画
-        opacity_anim = QPropertyAnimation(opacity, b"opacity")
-        opacity_anim.setDuration(actual_duration)
-        opacity_anim.setStartValue(0)
-        opacity_anim.setEndValue(1)
-
-        # 大小动画
-        size_anim = QPropertyAnimation(widget, b"size")
-        size_anim.setDuration(actual_duration)
-        size_anim.setStartValue(int(widget.width() * 0.9), int(widget.height() * 0.9))
-        size_anim.setEndValue(widget.width(), widget.height())
-        size_anim.setEasingCurve(QEasingCurve.Type.OutBack)
-
-        opacity_anim.start()
-        size_anim.start()
-
-    def scale_out(self, widget: QWidget, duration: int = 150):
-        """缩放退场"""
-        actual_duration = AnimationHelper._get_reduced_duration(duration)
-        if actual_duration < 10:
-            widget.hide()
-            return
-
-        opacity = QGraphicsOpacityEffect(widget)
-        widget.setGraphicsEffect(opacity)
-
-        # 透明度动画
-        opacity_anim = QPropertyAnimation(opacity, b"opacity")
-        opacity_anim.setDuration(actual_duration)
-        opacity_anim.setStartValue(1)
-        opacity_anim.setEndValue(0)
-
-        # 大小动画
-        size_anim = QPropertyAnimation(widget, b"size")
-        size_anim.setDuration(actual_duration)
-        size_anim.setStartValue(widget.width(), widget.height())
-        size_anim.setEndValue(int(widget.width() * 0.9), int(widget.height() * 0.9))
-        size_anim.setEasingCurve(QEasingCurve.Type.InBack)
-
-        opacity_anim.start()
-        size_anim.start()
-
-
-class AnimatedButton:
-    """带动画效果的按钮辅助类"""
-
-    @staticmethod
-    def add_ripple_effect(button, color: str = "rgba(255, 255, 255, 0.3)"):
-        """为按钮添加波纹效果（需要自定义绘制）"""
-        # 这是一个占位实现，实际波纹效果需要重写 paintEvent
-        button.setCursor(Qt.CursorShape.PointingHandCursor)
-
-    @staticmethod
-    def add_click_animation(button):
-        """为按钮添加点击动画
-
-        当系统启用了 prefers-reduced-motion 时：
-        - 跳过点击动画
-        """
-        button.setCursor(Qt.CursorShape.PointingHandCursor)
-        if AnimationHelper._prefers_reduced_motion():
-            return
-        button.clicked.connect(lambda: AnimatedButton._pulse_click(button))
-
-    @staticmethod
-    def _pulse_click(button):
-        """点击脉冲动画"""
-        original_style = button.styleSheet()
-
-        # 临时改变样式模拟按下效果
-        button.setStyleSheet(button.styleSheet() + "transform: scale(0.95);")
-
-        # 恢复
-        QTimer.singleShot(100, lambda: button.setStyleSheet(original_style))
-
-
-class LoadingAnimation:
-    """加载动画组件"""
-
-    @staticmethod
-    def create_loading_dots(parent, count: int = 3, color: str = "#6366F1"):
-        """创建加载点动画
-
-        当系统启用了 prefers-reduced-motion 时：
-        - 使用静态显示代替动画
-        """
-        from PySide6.QtWidgets import QWidget, QHBoxLayout, QLabel
-
-        container = QWidget(parent)
-        layout = QHBoxLayout(container)
-        layout.setSpacing(4)
-        layout.setContentsMargins(0, 0, 0, 0)
-
-        dots = []
-        for _i in range(count):
-            dot = QLabel("●")
-            dot.setStyleSheet(f"color: {color}; font-size: 10px;")
-            layout.addWidget(dot)
-            dots.append(dot)
-
-        # 根据减少动画偏好决定是否启用动画
-        if AnimationHelper._prefers_reduced_motion():
-            # reduce motion: 静态显示
-            for d in dots:
-                d.setStyleSheet(f"color: {color}; font-size: 10px; opacity: 1;")
-            return container, None
-
-        # 动画定时器
-        def animate():
-            for j, d in enumerate(dots):
-                delay = (_i + j) % count
-                opacity = 0.3 if delay != 0 else 1.0
-                d.setStyleSheet(f"color: {color}; font-size: 10px; opacity: {opacity};")
-
-        timer = QTimer(parent)
-        timer.timeout.connect(animate)
-        timer.start(200)
-
-        return container, timer
+__all__ = ["AnimationHelper"]
